Add MangaDetailActivity with two fragments, info and chapters

+++ app/src/main/java/eu/kanade/mangafeed/data/helpers/DatabaseHelper.java
@@ -98,7 +98,7 @@ public class DatabaseHelper implements MangaManager, ChapterManager {
     }
 
     @Override
-    public Observable<List<Manga>> getManga(int id) {
+    public Observable<List<Manga>> getManga(long id) {
         return mMangaManager.getManga(id);
     }
